Rename ServiceNaming.covert and move the package
