+++ core/src/main/java/com/linecorp/armeria/server/DefaultServiceConfigSetters.java
@@ -27,7 +27,6 @@ import javax.annotation.Nullable;
 
 import com.google.common.collect.ImmutableList;
 
-import com.linecorp.armeria.common.ServiceNaming;
 import com.linecorp.armeria.internal.server.annotation.AnnotatedService;
 import com.linecorp.armeria.server.logging.AccessLogWriter;
