+++ core/src/main/java/com/linecorp/armeria/server/VirtualHostServiceBindingBuilder.java
@@ -25,7 +25,6 @@ import java.util.function.Predicate;
 
 import com.linecorp.armeria.common.HttpMethod;
 import com.linecorp.armeria.common.MediaType;
-import com.linecorp.armeria.common.ServiceNaming;
 import com.linecorp.armeria.server.logging.AccessLogWriter;
 
 /**
